+++ planning/src/main/java/fape/core/planning/search/strategies/plans/LeastFlawRatio.java
@@ -24,17 +24,17 @@ public class LeastFlawRatio extends PartialPlanComparator {
     }
 
     @Override
-    public float g(State st) {
+    public double g(State st) {
         return 0;
     }
 
     @Override
-    public float h(State st) {
+    public double h(State st) {
         return eval(st);
     }
 
     @Override
-    public float hc(State st) {
+    public double hc(State st) {
         return eval(st);
     }
 }
